requirements.txt
langchain-openai
langchain-core
python-dotenv
streamlit
langchain-community
langserve
fastapi
uvicorn
sse_starlette
ipykernel
bs4
pypdf
chromadb
faiss-cpu
arxiv
wikipedia
langchainhub
groq
langchain-groq
sentence_transformers
langchain-huggingface
langchain-objectbox
langchain_google_genai
